Add files via upload
TCS (30 July, 2018)

This is the pipeline script used to calibrate the uid___A002_Xc6c0d5_X3f2e.ms file, which is one of the two datasets taken towards G328.25 in band 3 with the 7m array configuration. It contains the following modifications:

 - Additional manual flagging for the bandpass amplitude drops.
 - Additional Tsys flagging for the N2H+ line seen towards the target source.

The result of this calibration will be put in the database together with the corresponding weblog.