Create README_misc.md
I want to pass the README and scripts that I badly located in the upper level of the ALMA-IMF GitHub into a reduction/misc folder. 
I'm doing this on the web browser since I got a "There isn’t anything to compare" message doing things in terminal. I will figure that out later. This is good enough for now.