continuum_imaging_both.py
Turns out the crashes I had were not because of the presence of a PM antenna in one EB, but because of the format of the antenna lists for the "best sensitivity" imaging. There, several ms's are used, so several antenna lists are needed. The previous code concatenated all ms antenna lists to crate a single one, which then was passed to each of the ms's to image. So when, say, antenna "DA41" was not in some of the ms's the imaging crashed.