Update imaging_parameters.py
Edited default_lines. Changed some of its names to be more specific on the transitions. Names now are longer but we avoid confusions between, e.g., the two 13cs lines. I checked that all the lines for which I changed  the name have not been used in cubes (e.g., there are no more instances of the name in the script, other  than the definition of the line name). I noticed that he frequency for "ch3cn" might be wrong. Or at least I don't find anything close in Splatalogue. Several cubes have been made with that line frequency.  I also changed the threshold of "W43-MM2_B6_12M_robust0" back to 6 mJy, because the more specific value of 8.1 mJy is not used in "W43-MM2_B6_12M_robust0_sio".