Flagging channels for B6 spw5 JvM Cubes
First, I'm not sure of all the changes proposed. 

G012.80 : I made no changes, but the channels already flagged still have cleaning problems.

G333.60 : No divergence but the "pattern" seems present.

G351.77 : Weird "mosaic" behavior between 267 km/s and 277km/s and between -283km/s and -267km/s (not flagged because I don't know if it's a cleaning problem).
There are a lot of channels with important negative intensity areas --> is it a problem ?
The channels already flagged have no more divergences but still a weird pattern.

G353.41: I just extend a bit the flagged channels by precaution, but the flagged channels have still divergences

W43-MM2: No divergence, but there is the same pattern as before, and again a lot of negative intensity areas on a lot of channels.